+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/sync/backwardsync/BackwardsSyncAlgorithm.java
@@ -73,21 +73,22 @@ public class BackwardsSyncAlgorithm {
     if (!context.isReady()) {
       return waitForReady();
     }
-    runFinalizedSuccessionRule(
-        context.getProtocolContext().getBlockchain(), context.findMaybeFinalized());
-    final Optional<BlockHeader> possibleFirstAncestorHeader =
+
+    final Optional<BlockHeader> maybeFirstAncestorHeader =
         context.getBackwardChain().getFirstAncestorHeader();
-    if (possibleFirstAncestorHeader.isEmpty()) {
+    if (maybeFirstAncestorHeader.isEmpty()) {
       this.finished = true;
       LOG.info("The Backward sync is done");
       context.getBackwardChain().clear();
       return CompletableFuture.completedFuture(null);
     }
+
     final MutableBlockchain blockchain = context.getProtocolContext().getBlockchain();
-    final BlockHeader firstAncestorHeader = possibleFirstAncestorHeader.get();
+    final BlockHeader firstAncestorHeader = maybeFirstAncestorHeader.get();
     if (blockchain.contains(firstAncestorHeader.getHash())) {
       return executeProcessKnownAncestors();
     }
+
     if (blockchain.getChainHead().getHeight() > firstAncestorHeader.getNumber()) {
       debugLambda(
           LOG,
@@ -164,52 +165,6 @@ public class BackwardsSyncAlgorithm {
     if (context.isReady()) {
       latch.countDown();
     }
-  }
-
-  @VisibleForTesting
-  protected void runFinalizedSuccessionRule(
-      final MutableBlockchain blockchain, final Optional<Hash> maybeFinalized) {
-    if (maybeFinalized.isEmpty()) {
-      LOG.debug("Nothing to validate yet, consensus layer did not provide a new finalized block");
-      return;
-    }
-    final Hash newFinalized = maybeFinalized.get();
-    if (!blockchain.contains(newFinalized)) {
-      LOG.debug("New finalized block {} is not imported yet", newFinalized);
-      return;
-    }
-
-    final Optional<Hash> maybeOldFinalized = blockchain.getFinalized();
-    if (maybeOldFinalized.isPresent()) {
-      final Hash oldFinalized = maybeOldFinalized.get();
-      if (newFinalized.equals(oldFinalized)) {
-        LOG.debug("We already have this block as finalized");
-        return;
-      }
-      BlockHeader newFinalizedHeader =
-          blockchain
-              .getBlockHeader(newFinalized)
-              .orElseThrow(
-                  () ->
-                      new BackwardSyncException(
-                          "The header " + newFinalized.toHexString() + "not found"));
-      BlockHeader oldFinalizedHeader =
-          blockchain
-              .getBlockHeader(oldFinalized)
-              .orElseThrow(
-                  () ->
-                      new BackwardSyncException(
-                          "The header " + oldFinalized.toHexString() + "not found"));
-      LOG.info(
-          "Updating finalized {} block to new finalized block {}",
-          oldFinalizedHeader.toLogString(),
-          newFinalizedHeader.toLogString());
-    } else {
-      // Todo: should TTD test be here?
-      LOG.info("Setting new finalized block to {}", newFinalized);
-    }
-
-    blockchain.setFinalized(newFinalized);
   }
 
   private CompletableFuture<Void> waitForPeers(final int count) {